feat: client reservation n cancel reservation

+++ src/main/java/br/edu/ifsp/lp1p1/controller/BookController.java
@@ -2,6 +2,7 @@ package br.edu.ifsp.lp1p1.controller;
 
 import br.edu.ifsp.lp1p1.dto.book.BookResponseDTO;
 import br.edu.ifsp.lp1p1.dto.loan.LoanRequestDTO;
+import br.edu.ifsp.lp1p1.dto.reservation.ReservationRequestDTO;
 import br.edu.ifsp.lp1p1.mapper.book.BookResponseDTOMapper;
 import br.edu.ifsp.lp1p1.model.Book;
 import br.edu.ifsp.lp1p1.model.Loan;
@@ -62,6 +63,9 @@ public class BookController {
         Book book = this.bookService.findById(id);
         User client = this.userService.findById(loanRequestDTO.clientId());
         User user = this.userService.findByEmail(userDetails.getUsername());
+        if(book.getNumberOfCopiesAvailable() <= 0){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         book.setNumberOfCopiesAvailable(book.getNumberOfCopiesAvailable()-1);
         this.bookService.save(book);
         this.userService.save(client);
@@ -79,6 +83,43 @@ public class BookController {
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
 
+    //TODO????: LoanResponseDTO
+    @PostMapping("/{id}/reservation")
+    @Transactional
+    @PreAuthorize("hasAuthority('ADMIN') or hasAuthority('EMPLOYEE') or hasAuthority('CLIENT')")
+    public ResponseEntity<Loan> createReservation(@RequestBody ReservationRequestDTO reservationRequestDTO,
+                                                  @PathVariable Long id,
+                                                  @AuthenticationPrincipal UserDetails userDetails){
+        log.info(userDetails.getUsername());
+        log.info(userDetails.getAuthorities());
+        Book book = this.bookService.findById(id);
+        User client = this.userService.findById(this.userService.findByEmail(userDetails.getUsername()).getId());
+        User user = this.userService.findById(reservationRequestDTO.employeeId());
+        if(book.getNumberOfCopiesAvailable() <= 0){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        book.setNumberOfCopiesAvailable(book.getNumberOfCopiesAvailable()-1);
+        this.bookService.save(book);
+        this.userService.save(client);
+        this.userService.save(user);
+
+        if(reservationRequestDTO.daysAhead() > 3){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+
+        Loan loan = new Loan().builder()
+                .id(null)
+                .book(book)
+                .client(client)
+                .user(user)
+                .loanDate(LocalDateTime.now().plusDays(reservationRequestDTO.daysAhead()).truncatedTo(ChronoUnit.SECONDS).toInstant(ZoneOffset.UTC))
+                .returnDate(Instant.parse(reservationRequestDTO.returnDate())).build();
+
+        this.loanService.save(loan);
+        return new ResponseEntity<>(HttpStatus.CREATED);
+
+    }
+
     @GetMapping("/{id}/return")
     @Transactional
     @PreAuthorize("hasAuthority('ADMIN') or hasAuthority('EMPLOYEE')")
@@ -91,14 +132,41 @@ public class BookController {
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
         for(Loan l : loans){
-            if(l.getClient().getId() == clientId){
+            if(l.getClient().getId() == clientId && book.getId() == l.getBook().getId()){
                 this.loanService.deleteById(l.getId());
+                book.setNumberOfCopiesAvailable(book.getNumberOfCopiesAvailable()+1);
             }
         }
 
-        book.setNumberOfCopiesAvailable(book.getNumberOfCopiesAvailable()+1);
         this.bookService.save(book);
         BookResponseDTO bookResponseDTO = BookResponseDTOMapper.INSTANCE.toBookResponseDTO(book);
         return ResponseEntity.ok(bookResponseDTO);
     }
+
+    @GetMapping("/{id}/cancel/reservation")
+    @Transactional
+    @PreAuthorize("hasAuthority('ADMIN') or hasAuthority('EMPLOYEE') or hasAuthority('CLIENT')")
+    public ResponseEntity<BookResponseDTO> cancelReservation(@PathVariable Long id,
+                                                             @AuthenticationPrincipal UserDetails userDetails){
+
+        Book book = this.bookService.findById(id);
+
+        List<Loan> loans = this.loanService.findAllByClientId(this.userService.findByEmail(userDetails.getUsername()).getId());
+        if(loans.size() <= 0){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+
+        for(Loan l : loans){
+            if(l.getBook().getId() == id){
+                this.loanService.deleteById(l.getId());
+                book.setNumberOfCopiesAvailable(book.getNumberOfCopiesAvailable()+1);
+            }
+        }
+
+        this.bookService.save(book);
+        BookResponseDTO bookResponseDTO = BookResponseDTOMapper.INSTANCE.toBookResponseDTO(book);
+        return ResponseEntity.ok(bookResponseDTO);
+
+    }
+
 }
